+++ index.js
@@ -3,6 +3,7 @@ module.exports = {
     init: require('./lib/commands/init'),
     validate: require('./lib/commands/validate'),
     build: require('./lib/commands/build'),
-    studio: require('./lib/commands/studio')
+    test: require('./lib/commands/test'),
+    sandbox: require('./lib/commands/sandbox')
   }
 };